refactor(garm): move RunnerConfig into charm_state.py
RunnerConfig is relation-derived desired state, so per ISD014 (managing
charm complexity) it belongs in charm_state.py alongside CharmState and
SSHDebugInfo — not in the rendering module. runner_template.py now imports
it and only renders it; charm_state.py stays free of rendering/reconciler
imports to avoid a cycle.

Also hoist the large _OTEL_COLLECTOR_SETUP constant to the top of
runner_template.py with the other module constants, and document the
value-object placement rule in the garm AGENTS.md.
